Issue #11163: Enforce file size on InputParameterName and InputParameterNameOne
